% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.domain.features.R
\name{add.domain.features}
\alias{add.domain.features}
\title{Add information to the domain contacts entry}
\usage{
add.domain.features(contacts, resolution = 50000)
}
\arguments{
\item{contacts}{The list of domains generated from get.contacts()}

\item{resolution}{The size of the region at the start/end of the domain to define a corner}
}
\value{
The contacts list, with additional information
}
\description{
Add information to the domain contacts entry
}
