% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.domain.features.R
\name{check_order_2D_intervals}
\alias{check_order_2D_intervals}
\title{Check that the 2D intervals have interval1 is upstream of interval2}
\usage{
check_order_2D_intervals(intervals)
}
\description{
Check that the 2D intervals have interval1 is upstream of interval2
}
