% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_domains.R
\name{classify_domains.bak}
\alias{classify_domains.bak}
\title{Determine the group(s) to which a domain belongs}
\usage{
classify_domains.bak(domain_calls)
}
\arguments{
\item{domain_calls}{A data.frame with domains as rows and sectors as columns.}
}
\value{
Character vector of length nrow(domain_calls) with `_` separated enriched sectors.
}
\description{
A data.frame of domains and the calls of significant sectors is translated into the set of sectors that are enriched.
}
