% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_domains.R
\name{classify_domains.double}
\alias{classify_domains.double}
\title{Determine the group(s) to which a domain belongs}
\usage{
classify_domains.double(contacts, types, measures = c("P", "Q", "p", "Z"),
  thresholds, p.thresholds = c(0.0001, 0.00000001), z.quantiles = c(0.5,
  0.75), z.thresholds = NULL)
}
\arguments{
\item{contacts}{The list of domains generated from get_Q_values()}

\item{types}{The names of the elements to iterate over; these are produced by get.fend.samples() and are usually POTENTIAL (if get.fends was TRUE) and OBSERVED}

\item{measures}{The names of the p-value elements in Z_STATISTICS to iterate over; these are produced by get_Z_statistics()}

\item{thresholds}{A list of significance thresholds to apply}

\item{p.threshold}{The threshold to set for all p-value measures}

\item{z.quantile}{The quantile of Z scores above which are considered significant}

\item{z.threshold}{A threshold to apply, overriding z.quantile and ignoring `types`}
}
\value{
The contacts list, with additional information in the Z_STATISTICS
}
\description{
A threshold of significance is applied to the calculated p-values, and a domain classification attributed.
}
\details{
The thresholds list is a 2-level list, the first level indicates the null method used (eg 'OBSERVED') and the second level specifies the measure (eg 'Q' or 'Z').
}
