% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.contacts.R
\name{count.contacts}
\alias{count.contacts}
\title{Counts the number of extracted fend pairs}
\usage{
count.contacts(contacts)
}
\arguments{
\item{contacts}{The list of domains generated from get.contacts()}
}
\value{
The contacts list, with additional information
}
\description{
For every element of the CONTACTS list, count the number of fend pairs
}
