% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.domain.position.classifications.R
\name{get.domain.position.classifications}
\alias{get.domain.position.classifications}
\title{Get the values for sectors of each domain}
\usage{
get.domain.position.classifications(contacts)
}
\arguments{
\item{contacts}{The list of domains generated from get.contacts()}
}
\value{
The contacts list, with additional information
}
\description{
Generates a set of values that could be used to describe the distribution of fend pairs among the 4 domain sectors
}
