% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.high.score.fends.R
\name{get.high.score.fends}
\alias{get.high.score.fends}
\title{Subset high-scoring fend pairs from all observed contacts}
\usage{
get.high.score.fends(contacts, min_score = 50)
}
\arguments{
\item{contacts}{The list of domains generated from get.contacts() and add.domain.features()}

\item{min_score}{The minimum score that a fend pair must have to be retained}
}
\value{
The contacts list, with additional information
}
\description{
Select the fend pairs that are above a score threshold and save their information into the HIGH_SCORE slot of CONTACTS for each domain
}
