% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.contacts.R
\name{get_contacts}
\alias{get_contacts}
\title{Create a list of information for a set of intervals}
\usage{
get_contacts(domains,
  track_nm = "hic.Genome_Structure_Development.AST_NS_WT.tracks.observed_new_score",
  band = c(10000, 15000000),
  cache.path = "/home/cbarring/CLASSIFY_DOMAINS.CACHE",
  distribute_method = "PARALLEL", sge.resources = list(h_rt = "00:30:00"),
  max_jobs = 500, cluster_size = 10, get.fends = FALSE)
}
\arguments{
\item{domains}{A data.frame of 2D intervals}

\item{track_nm}{Name of the score track from which fend pairs are extracted}

\item{band}{Genomic distance between fend pairs to consider}

\item{cache.path}{Absolute path to the directory that cached domains can be found or saved}

\item{distribute_method}{Specify where the jobs to define each interval/domain should be distributed (using distributR)}

\item{sge.resources, max_jobs, cluster_size}{Parameters passed to distributR}

\item{get.fends}{Extracting the valid fragment ends and identifying potential 2D interactors (uses masses of memory!)}
}
\value{
The contacts list, with additional information
}
\description{
Extracts interaction frequency information for a set of 2D intervals. The domain information is cached
for quick loading, or loaded from the cache if the parameters have been run (and cached) before.
}
\seealso{
\code{distributR} for information on distributed jobs.
}
