% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Q_values.R
\name{get_Q_values}
\alias{get_Q_values}
\title{Correct for multiple testing in a data.frame}
\usage{
get_Q_values(contacts, types = c("POTENTIAL", "OBSERVED"), method = "BH")
}
\arguments{
\item{contacts}{The list of domains generated from get.contacts()}

\item{types}{Correct the p-values calculated from which null distribution(s)? `POTENTIAL` uses all theoretically possible fend pairs or `OBSERVED` uses the fend pairs that were identified.}

\item{method}{The correction method to apply, see p.adjust()}
}
\value{
The contacts list, with additional information in the Z_STATISTICS.
}
\description{
Uses p.adjust to correct for multiple testing on all p-values in the data.frame.
}
