% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Z_statistics.R
\name{get_Z_statistics}
\alias{get_Z_statistics}
\title{Calculates Z statistics and p-values}
\usage{
get_Z_statistics(contacts)
}
\arguments{
\item{contacts}{List of domain entries}
}
\value{
The contacts list, with additional information. Contains `P`, derived from the Z-statistic and `p` derived from the observed high-score count.
}
\description{
Calculates the probability that a domain sector contains more high-score fend pairs than expected by chance. Uses the
observed data and theoretical fend pairs as expected proportions.
}
\details{
Converts the Z-statistic to a p-value of enrichment using pnorm(-Z). For comparison, a p-value is
calculated using pnorm(OBS, mean=MEAN, sd=SDEV, lower.tail=FALSE). No correction for multiple testing is made.
}
