% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_domains.R
\name{load_domains}
\alias{load_domains}
\title{Load information on domain intervals}
\usage{
load_domains(dataset, project, band_from = 100000, band_to = 400000)
}
\arguments{
\item{project, dataset}{Name of the project and dataset that contain compaction tracks}

\item{band_from, band_to}{Size of the band used to define the compaction track}
}
\value{
A list with intervals for the CLUSTERS, DOMAINS and BORDERS
}
\description{
Load information on domain intervals
}
