% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_domains.R
\name{merge_domains}
\alias{merge_domains}
\title{Combine a list of domain contacts entries}
\usage{
merge_domains(contacts, intervalID = "merged")
}
\arguments{
\item{contacts}{List of domain contacts to be merged}

\item{intervalID}{Character string that will be used in the DOMAIN entry}
}
\value{
A list for a new meta-domain
}
\description{
Concatenates the fend pairs for ALL and HIGH_SCORE from a list of domains and
defines the interval as the region covered by the merged domains
}
\examples{
contacts <- append(contacts, merge_domains(contacts=contacts[1:3], intervalID='foo+bar+baz'))
print.fixed(plot_domain_contacts(contacts, ncol=3))

}
